Baseline commit
New project structure, separate views into separate files, move vehicle list API integration into a view model, fastlane test lanes, still leaving sample data in place for fuel entries, and vehicle list endpoint integration w/pagination
