+++ iOS Platform Assessment/Extension/Double+Extensions.swift
