+++ iOS Platform Assessment/Views/BrowseView.swift
