+++ iOS Platform Assessment/Views/FuelTypeSelector.swift
